backend/requirements.txt
fastapi
uvicorn
openai
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
